app2/requirements.txt
Flask
flask-cors
pymysql
requests
mysql-connector-python